Refactor tools and remove experimental flag
- Remove the experimental flag from the tools, making it a standard feature.
- Update README and CLI commands to reflect the removal of the `--exp` flag.
- Clean up the codebase by removing references to the experimental tools directory.
