+++ src/tlc_tools/cli/__init__.py
@@ -1,5 +1,7 @@
 """CLI package for tlc-tools."""
 
-from .registry import register_tool
+from .registry import register_tool  # noqa: I001
+from .commands.alias import main as alias_main  # noqa: F401, F403, I001
+from .commands.augment_bb_table import main as augment_bb_table_main  # noqa: F401, F403, I001
 
 __all__ = ["register_tool"]
